requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
supabase==2.7.4
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
websockets>=12.0
python-socketio>=5.10.0
python-engineio>=4.8.0
geopy>=2.4.0
requests>=2.31.0
httpx>=0.25.0
asyncpg>=0.29.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
